Refactor TargetFlowChecker to allow more tests
Currently the highlightTargetFlow method takes in two strings for the
source and destination annotations.  In order to allow for null-literal
checking to use the same graph traversal logic, the common
functionality has been moved to a separate method that takes two Query
objects and finds flows between them.  The highlightTargetFlow
functionality doesn't change, but it now just calls the more general
method.